+++ packages/ai-runtime/src/types/assistant/index.ts
@@ -1,3 +1,5 @@
+import type { RuntimeOutputVariable, RuntimeOutputVariablesSchema } from '../runtime';
+
 export * from './utils';
 export * from './yjs';
 
@@ -158,7 +160,7 @@ export type VariableType = VariableTypeBase &
 
 export type OutputVariable = VariableType & {
   variable?: { key: string; scope: string };
-  initialValue?: any;
+  initialValue?: RuntimeOutputVariablesSchema[RuntimeOutputVariable];
 };
 
 export interface Agent extends AssistantBase {